+++ components/mixinContainer.js
@@ -46,6 +46,7 @@ function mixinContainer(React) {
 
     componentDidMount: function () {
       this.registerStores(this.props)
+      if (this.props.onMount) this.props.onMount(this.props, this.context)
     },
 
     componentWillUnmount: function () {